Move helper functions to helpers/utils.py
