+++ music_assistant/providers/apple_music/helpers/__init__.py
@@ -1,5 +1,11 @@
 """Various Apple Music utils/helpers."""
 
 from .browse import browse_playlists
+from .utils import is_catalog_id, is_library_id, translate_media_type_to_apple_type
 
-__all__ = ["browse_playlists"]
+__all__ = [
+    "browse_playlists",
+    "is_catalog_id",
+    "is_library_id",
+    "translate_media_type_to_apple_type",
+]
